Address CodeRabbit review on read tools PR
- .github/copilot-instructions.md: document hcl/v2, go-cty, and x/sync
  as approved deps with rationale (HCL parser is the inverse of the
  renderer; errgroup has no stdlib equivalent for bounded fan-out).
- internal/tools: split read_helpers.go into team_source.go (fetch
  shape), read_errors.go (notConfigured/notFound), and as_json.go
  (JSON projection) so the 'no helpers.go' rule holds. asJSON now
  wraps json errors with fmt.Errorf per the errors-are-values rule.
- internal/tools/lookup_user.go: document datadog 'team' scope in the
  LookupUserMatch contract comment so callers see the real shape.
- internal/tools/read_tools_test.go: assert MemberCount and EnvCount
  for pt-corpus so regressions in uniqueCount/envCount fail loudly.
- docs/README.md: reference team_source.go (post-split filename).
